Add files via upload
This is a basic version of Transportation Management (Vehicle Insurance Management) System that's built using MySQL Workbench 8.0CE and Visual Studio Code. Current version needs to be optimized for better and modern look and feel which will be focused in the upcoming days but for now I am just publishing a current on the go of my project for reference.